contract-chatbot/scripts/requirements_processing.txt
PyMuPDF==1.23.26
pdfminer.six==20231228
typing-extensions==4.8.0 